requirements.txt
matplotlib
scipy
numpy
Pillow
bezier